+++ docs/source/releasenotes.rst
@@ -23,6 +23,7 @@ Upcoming release
   - Add source folder parameter to ``List Messages``
   - Add limit to ``Delete Messages``
   - Add keywords to add/remove labels from GMail messages
+  - Add keyword ``Do Message Actions`` for performing custom set of actions on selected messages
 
 6.7.2
 -----